+++ camera_link/__init__.py
@@ -1,0 +1,1 @@
+"""LAN camera prototype. Importing the package does not load OS/media backends."""